[#5] Return source code to original wespire format
At least one line of code was committed to the pg version of this file (ea8fbb3225e2cba28944bf781c9443d7a5b85159) that was missed when it was moved to its own repo. Instead of trying to backtrack, we'll return it to its original wespire state (which may be different than the original forked repo state), then re-commit each change here, for better documentation.
- copy entire source code from first wespire commit 4c0c7113730c3955f0c815f1ef5e196c4b2e3c65

[#5]